.history/app/chatgpt_ai/openia_20230331151354.py
from dotenv import load_dotenv
import open
import os 

load_dotenv()   

open.api_key = os.getenv("CHATGPT_API_KEY")


def chatgpt_response(prompt):
    response = openai.Completion.create(
        
    )