.history/requirements_20230331131910.txt
discord.py==2.1.0
opena
